pyproject.toml

[project]
name = "writing_toys"
version = "0.1.0"
description = "Some small tools used for idea generation in writing."
authors = [
    {name = "Matt Zhang", email = "set.stun@gmail.com"},
]
dependencies = [
    "py-gpt-interface>=0.3.2",
    "wonderwords>=2.2.0",
    "ai-emotion>=0.3.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

